+++ src/_igraph/convert.c
@@ -643,6 +643,31 @@ int igraphmodule_PyObject_to_subgraph_implementation_t(PyObject *o,
 
 /**
  * \ingroup python_interface_conversion
+ * \brief Converts a Python object to an igraph \c igraph_to_directed_t
+ */
+int igraphmodule_PyObject_to_to_directed_t(PyObject *o,
+  igraph_to_directed_t *result) {
+  static igraphmodule_enum_translation_table_entry_t to_directed_tt[] = {
+        {"acyclic",   IGRAPH_TO_DIRECTED_ACYCLIC},
+        {"arbitrary", IGRAPH_TO_DIRECTED_ARBITRARY},
+        {"mutual",    IGRAPH_TO_DIRECTED_MUTUAL},
+        {"random",    IGRAPH_TO_DIRECTED_RANDOM},
+        {0,0}
+  };
+
+  if (o == Py_True) {
+    *result = IGRAPH_TO_DIRECTED_MUTUAL;
+    return 0;
+  } else if (o == Py_False) {
+    *result = IGRAPH_TO_DIRECTED_ARBITRARY;
+    return 0;
+  }
+
+  return igraphmodule_PyObject_to_enum(o, to_directed_tt, (int*)result);
+}
+
+/**
+ * \ingroup python_interface_conversion
  * \brief Converts a Python object to an igraph \c igraph_to_undirected_t
  */
 int igraphmodule_PyObject_to_to_undirected_t(PyObject *o,